+++ packages/dataviews/src/components/dataviews-layouts/list/index.tsx
@@ -427,6 +427,8 @@ export default function ViewList< Item >( props: ViewListProps< Item > ) {
 		string | null | undefined
 	>( undefined );
 
+	const compositeRef = useRef< HTMLDivElement >( null );
+
 	// Update the active composite item when the selected item changes.
 	useEffect( () => {
 		if ( selectedItem ) {
@@ -465,7 +467,13 @@ export default function ViewList< Item >( props: ViewListProps< Item > ) {
 			const targetCompositeItemId = generateCompositeId( itemIdPrefix );
 
 			setActiveCompositeId( targetCompositeItemId );
-			document.getElementById( targetCompositeItemId )?.focus();
+			if (
+				compositeRef.current?.contains(
+					compositeRef.current.ownerDocument.activeElement
+				)
+			) {
+				document.getElementById( targetCompositeItemId )?.focus();
+			}
 		},
 		[ data, generateCompositeItemIdPrefix ]
 	);
@@ -536,6 +544,7 @@ export default function ViewList< Item >( props: ViewListProps< Item > ) {
 	if ( hasData && groupField && dataByGroup ) {
 		return (
 			<Composite
+				ref={ compositeRef }
 				id={ `${ baseId }` }
 				render={ <div /> }
 				className="dataviews-view-list__group"
@@ -601,6 +610,7 @@ export default function ViewList< Item >( props: ViewListProps< Item > ) {
 	return (
 		<>
 			<Composite
+				ref={ compositeRef }
 				id={ baseId }
 				render={ <div /> }
 				className={ clsx( 'dataviews-view-list', className, {